+++ src/components/CardList/CardList.styled.js
@@ -4,7 +4,7 @@ import { colors } from "../../styles/colors";
 export const ListTitleBox = styled.div`
   display: flex;
   justify-content: space-between;
-  margin: 2.8rem 0 0.9rem;
+  margin: 4rem 0 0.9rem;
   @media (max-width: 1200px) {
     flex-direction: column;
   }
