+++ src/components/ListButtonBox/ListButtonBox.styled.js
@@ -5,7 +5,7 @@ export const ListButtonBox = styled.div`
   display: flex;
   gap: 2rem;
   justify-content: center;
-  margin-top: 1.8rem;
+  margin-top: 2.5rem;
   padding: 1.5rem 0;
   align-items: center;
